Use `u8` instead of `u64` for verbosity field
It makes way more sense now that we have the `parse(from_occurrences)`
option in structopt.

Also updates the docs for `main!`.

+++ examples/verbosity_by_any_other_name_is_still_as_verbose.rs
@@ -6,7 +6,7 @@ use quicli::prelude::*;
 #[derive(Debug, StructOpt)]
 struct Cli {
     #[structopt(long = "verbose", short = "v", parse(from_occurrences))]
-    team_rockets_blasting_off_again: u64
+    team_rockets_blasting_off_again: u8,
 }
 
 main!(|cli_args: Cli, log_level: team_rockets_blasting_off_again| {
